requirements.txt
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
dnspython==2.0.0
idna==2.10
multidict==4.7.6
protobuf==3.13.0
pycparser==2.20
pymongo==3.11.0
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.1
python-decouple==3.3
python-telegram-bot==12.8
selenium==3.141.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
webwhatsapi==2.0.5
yarl==1.5.1
